Make TauHitObject abstract
