+++ osu.Game.Rulesets.Tau/Objects/Beat.cs
@@ -1,0 +1,6 @@
+namespace osu.Game.Rulesets.Tau.Objects
+{
+    public class Beat : TauHitObject
+    {
+    }
+}
