+++ osu.Game.Rulesets.Tau/Objects/Drawables/DrawableTauHitObject.cs
@@ -1,28 +1,17 @@
-﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
 using System;
-using System.Diagnostics;
-using osu.Framework.Graphics;
-using osu.Framework.Graphics.Shapes;
-using osu.Framework.Graphics.Containers;
-using osu.Framework.Input.Bindings;
 using osu.Game.Rulesets.Objects.Drawables;
-using osu.Game.Rulesets.Scoring;
-using osuTK;
-using osuTK.Graphics;
-using System.Linq;
-using osu.Framework.Allocation;
-using osu.Game.Rulesets.Tau.Configuration;
-using osu.Game.Rulesets.Tau.UI;
-using osu.Framework.Bindables;
 
 namespace osu.Game.Rulesets.Tau.Objects.Drawables
 {
-    public class DrawableTauHitObject : DrawableHitObject<TauHitObject>, IKeyBindingHandler<TauAction>
+    public abstract class DrawableTauHitObject : DrawableHitObject<TauHitObject>
     {
-        public Container Box;
-        public Container IntersectArea;
+        public DrawableTauHitObject(TauHitObject obj)
+            : base(obj)
+        {
+        }
 
         public Func<DrawableTauHitObject, bool> CheckValidation;
 
@@ -38,153 +27,8 @@ namespace osu.Game.Rulesets.Tau.Objects.Drawables
         /// <summary>
         /// The action that caused this <see cref="DrawableHit"/> to be hit.
         /// </summary>
-        public TauAction? HitAction { get; private set; }
+        public TauAction? HitAction { get; protected set; }
 
-        private bool validActionPressed;
-
-        protected sealed override double InitialLifetimeOffset => HitObject.TimePreempt;
-
-        public DrawableTauHitObject(TauHitObject hitObject)
-            : base(hitObject)
-        {
-            Anchor = Anchor.Centre;
-            Origin = Anchor.Centre;
-            RelativeSizeAxes = Axes.Both;
-            Size = Vector2.One;
-            AddRangeInternal(new Drawable[]
-                {
-                    Box = new Container
-                    {
-                        RelativePositionAxes = Axes.Both,
-                        Origin = Anchor.Centre,
-                        Anchor = Anchor.Centre,
-                        Alpha = 0.05f,
-                        Children = new Drawable[]{
-                            new Box{
-                                RelativeSizeAxes= Axes.Both
-                            },
-                            IntersectArea = new Container{
-                                Size = new Vector2(16),
-                                RelativeSizeAxes = Axes.None,
-                                Origin = Anchor.Centre,
-                                Anchor = Anchor.Centre,
-                                AlwaysPresent = true
-                            }
-                        }
-                    },
-                }
-            );
-
-            Rotation = hitObject.Angle;
-            Position = Vector2.Zero;
-        }
-
-        private readonly Bindable<float> size = new Bindable<float>(16); // Change as you see fit.
-
-        [BackgroundDependencyLoader(true)]
-        private void load(TauRulesetConfigManager config)
-        {
-            config?.BindWith(TauRulesetSettings.BeatSize, size);
-            size.BindValueChanged(value => Box.Size = new Vector2(value.NewValue), true);
-        }
-
-        protected override void UpdateInitialTransforms()
-        {
-            base.UpdateInitialTransforms();
-            var b = HitObject.Angle;
-            var a = b *= (float)(Math.PI / 180);
-
-            Box.FadeIn(HitObject.TimeFadeIn);
-            Box.MoveToY(-0.485f, HitObject.TimePreempt);
-        }
-
-        protected override void CheckForResult(bool userTriggered, double timeOffset)
-        {
-            Debug.Assert(HitObject.HitWindows != null);
-
-            if (CheckValidation == null) return;
-
-            if (!userTriggered)
-            {
-                if (!HitObject.HitWindows.CanBeHit(timeOffset))
-                    ApplyResult(r => r.Type = HitResult.Miss);
-
-                return;
-            }
-
-            if (CheckValidation.Invoke(this))
-            {
-                var result = HitObject.HitWindows.ResultFor(timeOffset);
-
-                if (result == HitResult.None)
-                    return;
-
-                if (!validActionPressed)
-                    ApplyResult(r => r.Type = HitResult.Miss);
-                else
-                    ApplyResult(r => r.Type = result);
-            }
-        }
-
-        protected override void UpdateStateTransforms(ArmedState state)
-        {
-            base.UpdateStateTransforms(state);
-
-            const double time_fade_hit = 250, time_fade_miss = 400;
-
-            switch (state)
-            {
-                case ArmedState.Idle:
-                    LifetimeStart = HitObject.StartTime - HitObject.TimePreempt;
-                    HitAction = null;
-
-                    break;
-
-                case ArmedState.Hit:
-                    var b = HitObject.Angle;
-                    var a = b *= (float)(Math.PI / 180);
-
-                    Box.ScaleTo(2f, time_fade_hit, Easing.OutCubic)
-                       .FadeColour(Color4.Yellow, time_fade_hit, Easing.OutCubic)
-                       .MoveToOffset(new Vector2(0, -.1f), time_fade_hit, Easing.OutCubic)
-                       .FadeOut(time_fade_hit);
-
-                    this.FadeOut(time_fade_hit);
-
-                    break;
-
-                case ArmedState.Miss:
-                    var c = HitObject.Angle;
-                    var d = c *= (float)(Math.PI / 180);
-
-                    Box.ScaleTo(0.5f, time_fade_miss, Easing.InCubic)
-                       .FadeColour(Color4.Red, time_fade_miss, Easing.OutQuint)
-                       .MoveToOffset(new Vector2(0, -.1f), time_fade_hit, Easing.OutCubic)
-                       .FadeOut(time_fade_miss);
-
-                    this.FadeOut(time_fade_miss);
-
-                    break;
-            }
-        }
-
-        public bool OnPressed(TauAction action)
-        {
-            if (Judged)
-                return false;
-
-            validActionPressed = HitActions.Contains(action);
-
-            var result = UpdateResult(true);
-
-            if (IsHit)
-                HitAction = action;
-
-            return result;
-        }
-
-        public void OnReleased(TauAction action)
-        {
-        }
+        protected override double InitialLifetimeOffset => HitObject.TimePreempt;
     }
 }
