+++ osu.Game.Rulesets.Tau/UI/TauDrawableRuleset.cs
@@ -28,7 +28,16 @@ namespace osu.Game.Rulesets.Tau.UI
 
         protected override ReplayInputHandler CreateReplayInputHandler(Replay replay) => new TauFramedReplayInputHandler(replay);
 
-        public override DrawableHitObject<TauHitObject> CreateDrawableRepresentation(TauHitObject h) => new DrawableTauHitObject(h);
+        public override DrawableHitObject<TauHitObject> CreateDrawableRepresentation(TauHitObject h)
+        {
+            switch (h)
+            {
+                case Beat beat:
+                    return new DrawableBeat(beat);
+                default:
+                    return null;
+            }
+        }
 
         protected override PassThroughInputManager CreateInputManager() => new TauInputManager(Ruleset?.RulesetInfo);
